move project files into module dir, add entrypoint method
